backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.3
psycopg2-binary==2.9.10
celery[redis]==5.4.0
redis==5.1.1
openai==1.51.0
jinja2==3.1.4
pydantic==2.9.2
pydantic-settings==2.5.2
sse-starlette==2.1.3
httpx==0.27.2
resend==2.5.0
PyJWT==2.9.0
